pyproject.toml
[tool.poetry]
name = "vbpr"
version = "0.1.0"
description = ""
authors = ["jchanxtarov <ryotaro.shimizu@zozo.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
numpy = "^1.21.2"
pandas = "^1.3.3"
torch = "^1.9.1"
torchvision = "0.12.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
timeout-decorator = "^0.5.0"
tqdm = "^4.62.3"
matplotlib = "^3.4.3"
Pillow = "^8.3.2"
requests = "^2.26.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
